refactor(checker): split type alias validation helpers
