+++ crates/tsz-checker/src/types/type_checking/mod.rs
@@ -34,6 +34,7 @@ mod type_alias_checking;
 mod type_alias_depth_helpers;
 mod type_alias_missing_name_coverage;
 mod type_alias_recursion_patterns;
+mod type_alias_variance;
 mod unused;
 
 /// Release the type-alias resolution scratch pool owned by this module tree.
